Refactor: 유저 프로필 이미지 컴포넌트 이동
